Moved CommandSwerveDrivetrain.java from the commands folder to the subsystems folder.  It's a subsystem Jim, not a command!
