+++ src/main/java/frc/robot/commands/Operator/ShooterModeShootWithPose.java
@@ -5,7 +5,7 @@
 package frc.robot.commands.Operator;
 
 import frc.robot.Constants;
-import frc.robot.commands.CommandSwerveDrivetrain;
+import frc.robot.subsystems.CommandSwerveDrivetrain;
 import frc.robot.subsystems.ShooterSubsystem;
 
 import com.ctre.phoenix6.mechanisms.swerve.SwerveRequest;
